requirements.txt
uvicorn==0.29.0
fastapi==0.110.2
pandas==2.2.2
python-dateutil==2.9.0